requirements.txt
django==4.2.1
djangorestframework==3.14.0
backports.zoneinfo==0.2.1;python_version<"3.9"
dj-database-url==0.5.0
django-lifecycle==1.0.0
python-decouple==3.6
pillow==9.2.0
gunicorn==20.1.0
psycopg2==2.9.3
django-widget-tweaks==1.4.12
django-filter==22.1
djangorestframework-camel-case==1.3.0
django-anymail==8.6
sentry-sdk==1.11.1
django-storages[google]
django-dbbackup==4.0.2
drf-yasg==1.21.5
django-import-export==3.1.0
django-admin-interface==0.24.2
django-tinymce==3.6.1